Rework allowed-character templates
Not entirely sure this is helpful -- it would probably also be
possible to do this on the Python side with regexes (although actually
the Unicode support in regexes might hurt us in this case).

Also, while this is architecture-astronauting a bit, it is still much
saner than the previous implementation.