Revert re_compile to original behavior
This undoes some of the change from
e68757374266d36fef4b64b257dd342a474a22cf:

Fabian argued that a function that is documented as doing the same thing
as re.compile, except with a permanent cache actually should behave that
way, instead of failing in sys.exit etc.

Also, when running with checkpatterns, save the error message from re.